Create new folder structure for services and move scoring service there. Convert GET to POST and create basic scoring config.
