+++ package.json
@@ -45,6 +45,7 @@
     "hapi-pino": "^12.1.0",
     "hapi-pulse": "^3.0.1",
     "https-proxy-agent": "^7.0.5",
+    "joi": "^17.13.3",
     "lodash": "^4.17.21",
     "node-fetch": "^3.3.2",
     "pino": "^9.5.0",
