+++ src/api/scoring/index.js
@@ -1,3 +1,4 @@
+import Joi from 'joi'
 import { scoringController } from '~/src/api/scoring/controller.js'
 
 /**
@@ -8,9 +9,25 @@ const scoring = {
     name: 'scoring',
     register: (server) => {
       server.route({
-        method: 'GET',
-        path: '/scoring',
-        ...scoringController
+        method: 'POST',
+        path: '/score',
+        ...scoringController,
+        options: {
+          validate: {
+            payload: Joi.object({
+              answers: Joi.array()
+                .items(
+                  Joi.object({
+                    questionId: Joi.string().required(),
+                    answer: Joi.alternatives()
+                      .try(Joi.string(), Joi.number())
+                      .required()
+                  })
+                )
+                .required()
+            })
+          }
+        }
       })
     }
   }
