+++ src/config/scoring-data.js
@@ -1,0 +1,10 @@
+export const scoringData = [
+  {
+    id: 'q1',
+    answers: [
+      { answer: 'Protected cropping', score: 3 },
+      { answer: 'Fruit', score: 2 },
+      { answer: 'Field-scale crops', score: 1 }
+    ]
+  }
+]
