+++ src/services/scoring/simple-score.js
