+++ src/services/scoring/simple-score.test.js
